Divide traits from implementation
