+++ src/game_loop/html_game_loop.rs
@@ -1,15 +1,6 @@
-mod html_game_loop;
-use crate::{game_state::GameState, key_event::KeyEvent};
+use crate::{game_loop::GameLoop, game_state::GameState, key_event::KeyEvent};
 use std::cell::RefCell;
 use std::rc::Rc;
-
-pub trait GameLoop<T, U>
-where
-    T: GameState<U>,
-    U: KeyEvent,
-{
-    fn run(game_state: Rc<RefCell<T>>, key_event: Rc<RefCell<U>>);
-}
 
 pub struct HtmlGameLoop;
